// src/app/types/enum/http-header-key.enum.ts
export enum HttpHeaderKey {
    AUTHORIZATION = 'Authorization',
    CONTENT_TYPE = 'Content-Type',
    X_LOCALE = 'X-locale',
    X_CLIENTID = 'X-clientid',
    ACCEPT_LANGUAGE = 'Accept-Language',
    TX_ID = 'txid',
    FACEBOOK_TOKEN = 'facebookToken',
    GOOGLE_TOKEN = 'googleToken',
    UID = 'uid',
    UID_TYPE = 'uidType',
    BEARER = 'Bearer'
}